chore: move building related pages into Buildings folder
