+++ NextGen-BM-FE/src/pages/Buildings/BuildingPage.tsx
